Added Assembly Option to Output tar.gz
Issue 1556 expressed interest in using the linuxserver.io docker
base. This provides additional build and run options allowing
control of the ACLs of the filesystem. Having the build in a
tar.gz allows for better control of the extractions
`--strip-components=1` to remove the base directory name that
changes with each release. ZIP options only allow to flatten all
directories to 1 and lose the existing structure. Using the tar
allows the continused use of `COPY /root/ /` in Dockerfile and
keep the copying simpler.